Code reorganized according to comments
IObject2D moded to Common.Interface
Object2DProperties moved to Common.Properties
CompoundLayer moved to Common.Properties
