+++ Architecture_oM/Architecture_oM.csproj
@@ -41,12 +41,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Elements\CompoundLayer.cs" />
     <Compile Include="Elements\Grid.cs" />
     <Compile Include="Elements\Roof.cs" />
-    <Compile Include="Properties\Object2DProperties.cs" />
     <Compile Include="Elements\Wall.cs" />
-    <Compile Include="Elements\IObject2D.cs" />
     <Compile Include="Elements\Level.cs" />
     <Compile Include="Elements\Floor.cs" />
     <Compile Include="_Properties\AssemblyInfo.cs" />
@@ -65,6 +62,8 @@
       <Name>Geometry_oM</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Properties\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
